Add files via upload
Veri-Charm_demo.html is the fusion of all 3 web app components to simulate the user/buyer experience in authentication of purchased products in a trustless and permissionless manner. The range of applications in production/generation/manufacturing industries is quite extensive: agriculture, cosmetics and clothing, sportswear, electronics, pharmaceuticals, etc.

However, our primary focus will be how the Veri-Charm Protocol can be used in conjunction with the Bit-Earth Labs initiatives of BTC-PIN, BTC-Sci and the Bit-Earth Movement for a third party approvals-and-guarantees free execution environment while pursuing the Bitcoin-Energy Standard in a trustless manner.